# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OMEDIA",
    "description": "Disk-first media organization API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/rules/references": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Rule References",
        "operationId": "rule_references",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleReferences"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/organize": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "List Organize Rules",
        "operationId": "list_organize_rules",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrganizeRule"
                  },
                  "type": "array",
                  "title": "Response List Organize Rules"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Create Organize Rule",
        "operationId": "create_organize_rule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizeRuleDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizeRule"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/organize/preview": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Preview Organize Rule",
        "operationId": "preview_organize_rule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizeRulePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulePreview"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/organize/{rule_id}": {
      "put": {
        "tags": [
          "rules"
        ],
        "summary": "Save Organize Rule",
        "operationId": "save_organize_rule",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizeRuleDraft"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizeRule"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "rules"
        ],
        "summary": "Delete Organize Rule",
        "operationId": "delete_organize_rule",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/transfer": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "List Transfer Rules",
        "operationId": "list_transfer_rules",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransferRule"
                  },
                  "type": "array",
                  "title": "Response List Transfer Rules"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Create Transfer Rule",
        "operationId": "create_transfer_rule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRuleDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferRule"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/transfer/preview": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Preview Transfer Rule",
        "operationId": "preview_transfer_rule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRulePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulePreview"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/rules/transfer/{rule_id}": {
      "put": {
        "tags": [
          "rules"
        ],
        "summary": "Save Transfer Rule",
        "operationId": "save_transfer_rule",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferRuleDraft"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferRule"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "rules"
        ],
        "summary": "Delete Transfer Rule",
        "operationId": "delete_transfer_rule",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorDetail"
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse"
      },
...
      "OrganizeRule": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/RuleCategory"
            },
            "type": "array",
            "title": "Categories"
          },
          "fallback_bucket": {
            "type": "string",
            "title": "Fallback Bucket",
            "default": ""
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "OrganizeRule"
      },
      "OrganizeRuleDraft": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/RuleCategory"
            },
            "type": "array",
            "title": "Categories"
          },
          "fallback_bucket": {
            "type": "string",
            "title": "Fallback Bucket",
            "default": ""
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "OrganizeRuleDraft"
      },
      "OrganizeRulePreviewRequest": {
        "properties": {
          "rule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrganizeRuleDraft"
              },
              {
                "$ref": "#/components/schemas/OrganizeRule"
              }
            ],
            "title": "Rule"
          },
          "relative_path": {
            "type": "string",
            "format": "path",
            "title": "Relative Path"
          },
          "tmdb": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tmdb"
          }
        },
        "type": "object",
        "required": [
          "rule",
          "relative_path"
        ],
        "title": "OrganizeRulePreviewRequest"
      },
...
      "RulePreview": {
        "properties": {
          "bucket": {
            "type": "string",
            "title": "Bucket"
          },
          "matched_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Category"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "variables": {
            "additionalProperties": true,
            "type": "object",
            "title": "Variables"
          },
          "transfer_bucket": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transfer Bucket"
          },
          "organize_prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organize Prefix"
          },
          "media_relative_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Relative Path"
          },
          "destination_relative_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Relative Path"
          },
          "path_split_confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Split Confidence"
          }
        },
        "type": "object",
        "required": [
          "bucket"
        ],
        "title": "RulePreview"
      },
...
      "RuleReferenceGenre": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "name_zh": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name Zh"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "RuleReferenceGenre"
      },
      "RuleReferences": {
        "properties": {
          "movie_genres": {
            "items": {
              "$ref": "#/components/schemas/RuleReferenceGenre"
            },
            "type": "array",
            "title": "Movie Genres"
          },
          "tv_genres": {
            "items": {
              "$ref": "#/components/schemas/RuleReferenceGenre"
            },
            "type": "array",
            "title": "Tv Genres"
          },
          "countries": {
            "items": {
              "$ref": "#/components/schemas/RuleReferenceCountry"
            },
            "type": "array",
            "title": "Countries"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "bundled"
          }
        },
        "type": "object",
        "required": [
          "countries"
        ],
        "title": "RuleReferences"
      },
...
      "TransferRule": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/RuleCategory"
            },
            "type": "array",
            "title": "Categories"
          },
          "fallback_bucket": {
            "type": "string",
            "title": "Fallback Bucket",
            "default": ""
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "TransferRule"
      },
      "TransferRuleDraft": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/RuleCategory"
            },
            "type": "array",
            "title": "Categories"
          },
          "fallback_bucket": {
            "type": "string",
            "title": "Fallback Bucket",
            "default": ""
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TransferRuleDraft"
      },
      "TransferRulePreviewRequest": {
        "properties": {
          "rule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransferRuleDraft"
              },
              {
                "$ref": "#/components/schemas/TransferRule"
              }
            ],
            "title": "Rule"
          },
          "relative_path": {
            "type": "string",
            "format": "path",
            "title": "Relative Path"
          }
        },
        "type": "object",
        "required": [
          "rule",
          "relative_path"
        ],
        "title": "TransferRulePreviewRequest"
      },
...
    }
  }
}